requirements.txt
vk-api==11.8.0
dialogflow==1.0.0
python-telegram-bot==11.1.0
python-dotenv==0.14.0